+++ .github/workflows/api-ci-workflow.yml
@@ -1,16 +1,16 @@
-name: CI workflows
+name: API CI Workflow
 
 on:
   push:
     branches:
       - "*"
     paths:
-      - "app/src/**/*"
+      - "api/src/**/*"
   pull_request:
     branches:
       - "*"
     paths:
-      - "app/src/**/*"
+      - "api/src/**/*"
 
 jobs:
   building-stage:
@@ -18,41 +18,37 @@ jobs:
 
     steps:
       - name: Checkout repository
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
+      - name: Set up Eclipse Temurin Java
+        uses: actions/setup-java@v4
         with:
-          node-version: "18"
+          distribution: "temurin"
+          java-version: "21"
 
-      - name: Cache node modules
+      - name: Cache SBT dependencies
         uses: actions/cache@v3
         with:
           path: |
-            ~/.yarn
-            **/node_modules
-          key: ${{ runner.os }}-node-${{ hashFiles('**/yarn.lock') }}
+            ~/ .ivy2/cache
+            ~/ .sbt
+            ~/ .m2
+          key: ${{ runner.os }}-sbt-${{ hashFiles('**/build.sbt') }}
           restore-keys: |
-            ${{ runner.os }}-node-
+            ${{ runner.os }}-sbt-
 
-      - name: Install dependencies
-        working-directory: ./app
-        run: yarn install
+      - name: Run ScalaStyle
+        working-directory: ./api
+        run: sbt scalastyle
 
-      - name: Run linter
-        working-directory: ./app
-        run: yarn lint
+      - name: Run Muti Test
+        working-directory: ./api
+        run: sbt test
 
-      - name: Build Docekr image
-        working-directory: ./app
-        run: docker build . -t app
-  # notify-slack:
-  #   if: always()
-  #   needs: building-stage
-  #   uses: ./.github/workflows/slack-notification.yml
-  #   with:
-  #     f-slack-message: CI Failed! Please confirm.
-  #     s-slack-message: CI Completion! Good job.
+      - name: Build the Docker image
+        working-directory: ./api
+        run: docker build -t api .
+
   slack-notify:
     runs-on: ubuntu-latest
     steps:
@@ -60,7 +56,7 @@ jobs:
         uses: rtCamp/action-slack-notify@v2
         if: ${{ success() }}
         env:
-          SLACK_USERNAME: CI Github Actions
+          SLACK_USERNAME: API CI Github Actions
           SLACK_ICON: https://github.githubassets.com/images/modules/logos_page/GitHub-Mark.png
           SLACK_WEBHOOK: ${{ secrets.SLACK_WEBHOOK_URL }}
           SLACK_TITLE: Successfully🎉
@@ -70,7 +66,7 @@ jobs:
         uses: rtCamp/action-slack-notify@v2
         if: ${{ failure() }}
         env:
-          SLACK_USERNAME: CI Github Actions
+          SLACK_USERNAME: API CI Github Actions
           SLACK_ICON: https://github.githubassets.com/images/modules/logos_page/GitHub-Mark.png
           SLACK_WEBHOOK: ${{ secrets.SLACK_WEBHOOK_URL }}
           SLACK_TITLE: Failurefully⚠️
