+++ api/src/main/scala/com/art/collection/adapters/web/handler/Health.scala
@@ -1,13 +1,12 @@
-package adapters.web.handler
+package com.art.collection.adapters.web.handler
 
 import cats.effect.kernel.{Async}
-import org.http4s.dsl.impl./
 import org.http4s.dsl.io.*
 import org.http4s.{HttpRoutes, Response, Status}
 
 object Health {
   def healthHandler[F[_]: Async] = HttpRoutes.of[F] {
-    case GET -> Root / "health" =>
+    case GET -> Root =>
       Async[F].pure(Response(Status.Ok).withEntity("ok"))
   }
 }
