+++ api/src/main/scala/com/art/collection/adapters/web/routes/Route.scala
@@ -1,7 +1,7 @@
-package adapters.web.routes
+package com.art.collection.adapters.web.routes
 
-import adapters.web.handler.Health.healthHandler
 import cats.effect.kernel.Async
+import com.art.collection.adapters.web.handler.Health.healthHandler
 import org.http4s.HttpApp
 import org.http4s.server.Router
